Move AccessTuple out of the API, into the example
The AccessTuple type limits the API to a token database being stored in
a single process, in memory.

We remove this limit by moving the AccessTuple out of the git-oauth2
package into the example. The ginoauth2.Auth function now only needs an
accessCheckFunction closure that will authenticate users.
The basic authentication hard coded is now implemented only in the
example.