Update README.md
"Specify your chosen credentials and scopes for authorization, and redirect the URL to get redirected to after the Google OAuth2 was done, a secret for the CookieStore and a session name to be used in [Sessions as identifier](https://github.com/gin-gonic/contrib/sessions)."  -- this doesn't quite flow. Can you rewrite a bit?

https://github.com/gin-gonic/contrib/sessions is broken 

This isn't clear -- can you reword? "Created your Google clientid stored in example/test-clientid.google.json.hide similar to and grant redirect to example/test-clientid.google.json as described in the [article](http://skarlso.github.io/2016/06/12/google-signin-with-go/) mentioned before."